Begin moving roles and routes to own file
